requirements.txt
discord.py==1.2.3
PyNaCl==1.3.0
requests==2.22.0
youtube_dl==2019.9.12.1
